requirement.txt
Django==5.1.7
django-taggit==6.1.0
Pillow==11.1.0
psycopg2==2.9.10
djangorestframework==3.15.2
django-filter==25.1
drf-nested-routers==0.94.1